+++ server/src/tests.rs
@@ -6405,3 +6405,4 @@ mod encounters;
 mod heavy_turret;
 mod m01;
 mod pellets;
+mod shiv;